Add puzzle 11, part 1 "solution"
I commmitted this as failing at first, but only as I'd understood the
output wrong. This doesn't calculate the final result; it just tells you
where you end up on my representation of a hex grid.

If you look at a hex such that the tops and bottoms of the cells are
horizontal, and then take the zig-zagging verticals and flatten them,
you get a grid that's simply offset up or down each row:
  ____        ____        ____       ______        ______        ______
 ╱    ╲      ╱    ╲      ╱    ╲     |      |      |      |      |      |
╱ x -2 ╲____╱ x  0 ╲____╱ x  2 ╲    | x -2 |______| x  0 |______| x  2 |
╲ y  2 ╱    ╲ y  2 ╱    ╲ y  2 ╱    | y  2 |      | y  2 |      | y  2 |
 ╲____╱ x -1 ╲____╱ x  1 ╲____╱     |______| x -1 |______| x  1 |______|
 ╱    ╲ y  1 ╱    ╲ y  1 ╱    ╲     |      | y  1 |      | y  1 |      |
╱ x -2 ╲____╱ x  0 ╲____╱ x  2 ╲ == | x -2 |______| x  0 |______| x  2 |
╲ y  0 ╱    ╲ y  0 ╱    ╲ y  0 ╱ == | y  0 |      | y  0 |      | y  0 |
 ╲____╱ x -1 ╲____╱ x  1 ╲____╱     |______| x -1 |______| x  1 |______|
 ╱    ╲ y -1 ╱    ╲ y -1 ╱    ╲     |      | y -1 |      | y -1 |      |
╱ x -2 ╲____╱ x  0 ╲____╱ x  2 ╲    | x -2 |______| x  0 |______| x  2 |
╲ y -2 ╱    ╲ y -2 ╱    ╲ y -2 ╱    | y -2 |      | y -2 |      | y -2 |
 ╲____╱      ╲____╱      ╲____╱     |______|      |______|      |______|

Now the taxicab distance is just the absolute of the coordinates, then
whatever amount they share in common (subtract the lowest from both,
keep it) and then, sum in the remainder, dividing by 2 of it's in Y.